requirements.txt
easyocr
numpy
opencv_python
opencv_python_headless
pillow
streamlit
ultralytics
